backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
anthropic
python-multipart
aiofiles
jinja2
python-dotenv
